Домашнее задание 5 | Поправил некоторые типы данных в Fact и Part, потому что выпадала ошибка с неправильно принятыми данными. Сделал рефакторинг кода во всех файлах ?? Не понимаю почему у меня не обновляются показатели погоды, в одном городе запрашивается и потом дублируется во всех остальных городах, возможно пробелмы у Яндекса ? ((

+++ app/src/main/java/com/example/kotlincourseappivanov/model/dto/Fact.kt
@@ -33,5 +33,5 @@ data class Fact(
     @SerializedName("wind_gust")
     val windGust: Double,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
